Refactor error handling for low-level clients
Main changes:

* Allow event handler to return parsing errors and pass them to the
error handler. This allows to simplify event handler by returning errors
and handling them once in error handler instead if handling them in two
places.
* Replace boolean return argument in error handler with error. It sill
can control if client should reconnect or close with nil and non nil
values. In addition to that it allows to pass errors from the error
handler back to the function that started the stream. Thus Start()
function signature is also extended to allow returning error.